Expose the `--raw-port` flag in the run and run-prepared subcommands
Fixes #2113
